improve: merge to single file
