+++ packages/logger/tsconfig.json
@@ -3,5 +3,5 @@
 	"compilerOptions": {
 		"outDir": "dist"
 	},
-	"include": ["src/**/*.ts", "src/**/*.js"]
+	"include": ["src/**/*.ts"]
 }
